requirements.txt
aiogram
huggingface_hub
python-dotenv   
transformers
torch

